make the test conditional for >=go1.13
